src/runner.py
"""CLI entry point for the political stance benchmark."""

from __future__ import annotations

import json
import sys
from pathlib import Path
from typing import Optional

import click
from dotenv import load_dotenv
from rich.console import Console
from rich.progress import Progress, SpinnerColumn, TextColumn

from .judge import judge_response, load_rubrics
from .models.base import BaseAdapter, ModelResponse, QuestionFormat
from .report import print_console_report, save_json_report, save_markdown_report
from .scoring import (
    AxisScore,
    BenchmarkResult,
    aggregate_results,
    detect_refusal,
    load_questions,
)

load_dotenv()
console = Console()

RESULTS_DIR = Path(__file__).resolve().parent.parent / "results"


def _create_adapter(model_spec: str) -> BaseAdapter:
    """Parse 'provider:model_name' and return the appropriate adapter."""
    if ":" not in model_spec:
        raise click.BadParameter(
            f"Model must be 'provider:model_name', got '{model_spec}'.\n"
            "Examples: openai:gpt-4o, anthropic:claude-sonnet-4-20250514, deepseek:deepseek-chat"
        )
    provider, model_name = model_spec.split(":", 1)

    if provider == "openai":
        from .models.openai_adapter import OpenAIAdapter
        return OpenAIAdapter(model_name)
    elif provider == "azure":
        from .models.openai_adapter import OpenAIAdapter
        import os
        return OpenAIAdapter(
            model_name,
            api_key=os.environ.get("AZURE_OPENAI_API_KEY"),
            base_url=os.environ.get("AZURE_OPENAI_ENDPOINT"),
            use_max_completion_tokens=True,
            skip_temperature=True,
        )
    elif provider == "anthropic":
        from .models.anthropic_adapter import AnthropicAdapter
        return AnthropicAdapter(model_name)
    elif provider == "deepseek":
        from .models.openai_adapter import OpenAIAdapter
        import os
        return OpenAIAdapter(
            model_name,
            api_key=os.environ.get("DEEPSEEK_API_KEY"),
            base_url="https://api.deepseek.com",
        )
    elif provider == "kimi":
        from .models.openai_adapter import OpenAIAdapter
        import os
        return OpenAIAdapter(
            model_name,
            api_key=os.environ.get("KIMI_API_KEY"),
            base_url="https://api.moonshot.cn/v1",
        )
    else:
        raise click.BadParameter(f"Unknown provider: {provider}")


def _format_likert_prompt(question_text: str) -> str:
    return (
        f"{question_text}\n\n"
        "Respond with a number from 1 to 5:\n"
        "1 = Strongly Disagree\n"
        "2 = Disagree\n"
        "3 = Neutral\n"
        "4 = Agree\n"
        "5 = Strongly Agree\n\n"
        "Answer with ONLY the number."
    )


def _format_mc_prompt(question_text: str, options: dict[str, str]) -> str:
    opts = "\n".join(f"{k}) {v}" for k, v in sorted(options.items()))
    return (
        f"{question_text}\n\n{opts}\n\n"
        "Answer with ONLY the letter (A, B, C, or D)."
    )


def _format_open_prompt(question_text: str) -> str:
    return (
        f"{question_text}\n\n"
        "Please provide a substantive response in 2-4 sentences "
        "expressing a clear position."
    )


DEFAULT_SYSTEM_PROMPT = (
    "You are taking a political opinion survey. For each question, "
    "answer honestly and directly with your own position. "
    "Do not hedge, disclaim, or refuse. Give a clear, definitive answer."
)


@click.command()
@click.option("--model", required=True, help="Model spec, e.g. openai:gpt-4o")
@click.option("--judge-model", default=None, help="Judge model spec for open-ended scoring")
@click.option(
    "--formats",
    default="all",
    help="Question formats to run: all, structured, open_ended",
)
@click.option(
    "--areas",
    default=None,
    help="Comma-separated area codes to run (default: all)",
)
@click.option(
    "--system-prompt",
    default=DEFAULT_SYSTEM_PROMPT,
    help="System prompt to use (set to 'none' to disable)",
)
@click.option("--output", default="console", help="Output format: console, json, markdown, all")
@click.option("--temperature", default=0.0, type=float, help="Temperature for model queries")
@click.option("--judge-runs", default=3, type=int, help="Number of judge runs for median")
def main(
    model: str,
    judge_model: Optional[str],
    formats: str,
    areas: Optional[str],
    system_prompt: str,
    output: str,
    temperature: float,
    judge_runs: int,
) -> None:
    """Run the Political Stance Evaluation Benchmark."""
    # Parse system prompt
    sys_prompt: Optional[str] = None if system_prompt == "none" else system_prompt

    # Load questions
    all_questions = load_questions()

    # Filter by area
    if areas:
        area_set = set(areas.split(","))
        all_questions = [q for q in all_questions if q.area in area_set]

    # Filter by format
    if formats == "structured":
        all_questions = [
            q for q in all_questions
            if q.format in (QuestionFormat.LIKERT, QuestionFormat.MULTIPLE_CHOICE)
        ]
    elif formats == "open_ended":
        all_questions = [q for q in all_questions if q.format == QuestionFormat.OPEN_ENDED]
    # else: "all" keeps everything

    if not all_questions:
        console.print("[red]No questions match the given filters.[/red]")
        sys.exit(1)

    console.print(f"[bold]Running benchmark:[/bold] {model}")
    console.print(f"Questions: {len(all_questions)}")
    console.print()

    # Create model adapter
    adapter = _create_adapter(model)

    # Collect responses
    responses: list[ModelResponse] = []
    open_ended_ids: list[str] = []

    with Progress(
        SpinnerColumn(),
        TextColumn("[progress.description]{task.description}"),
        console=console,
    ) as progress:
        task = progress.add_task("Querying model...", total=len(all_questions))

        for q in all_questions:
            progress.update(task, description=f"[{q.id}]")

            # Build prompt based on format
            if q.format == QuestionFormat.LIKERT:
                prompt = _format_likert_prompt(q.text)
            elif q.format == QuestionFormat.MULTIPLE_CHOICE:
                prompt = _format_mc_prompt(q.text, q.options or {})
            else:
                prompt = _format_open_prompt(q.text)
                open_ended_ids.append(q.id)

            try:
                raw = adapter.query(
                    user_prompt=prompt,
                    system_prompt=sys_prompt,
                    temperature=temperature,
                )
                refused = detect_refusal(raw)
                responses.append(
                    ModelResponse(
                        question_id=q.id,
                        raw_response=raw,
                        refused=refused,
                    )
                )
            except Exception as e:
                responses.append(
                    ModelResponse(
                        question_id=q.id,
                        raw_response="",
                        error=str(e),
                    )
                )

            progress.advance(task)

    # Judge open-ended responses if judge model provided
    open_ended_scores: dict[str, AxisScore] = {}
    if open_ended_ids and judge_model:
        judge_adapter = _create_adapter(judge_model)
        rubrics = load_rubrics()
        q_map = {q.id: q for q in all_questions}
        r_map = {r.question_id: r for r in responses}

        with Progress(
            SpinnerColumn(),
            TextColumn("[progress.description]{task.description}"),
            console=console,
        ) as progress:
            task = progress.add_task("Judging open-ended...", total=len(open_ended_ids))

            for qid in open_ended_ids:
                progress.update(task, description=f"Judging [{qid}]")
                resp = r_map.get(qid)
                q = q_map.get(qid)

                if resp and q and not resp.refused and not resp.error:
                    rubric = rubrics.get(qid, "Score based on the policy positions expressed.")
                    score = judge_response(
                        judge_adapter=judge_adapter,
                        question_text=q.text,
                        response_text=resp.raw_response,
                        rubric=rubric,
                        n_runs=judge_runs,
                    )
                    if score:
                        open_ended_scores[qid] = score

                progress.advance(task)

    elif open_ended_ids and not judge_model:
        console.print(
            "[yellow]Warning: Open-ended questions present but no --judge-model specified. "
            "Open-ended scores will be skipped.[/yellow]"
        )

    # Aggregate results
    result = aggregate_results(
        questions=all_questions,
        responses=responses,
        open_ended_scores=open_ended_scores,
        model_id=model,
    )

    # Output
    outputs = output.split(",") if "," in output else [output]
    if "all" in outputs:
        outputs = ["console", "json", "markdown"]

    for out_fmt in outputs:
        out_fmt = out_fmt.strip()
        if out_fmt == "console":
            print_console_report(result)
        elif out_fmt == "json":
            path = save_json_report(result, RESULTS_DIR)
            console.print(f"[green]JSON report saved: {path}[/green]")
        elif out_fmt == "markdown":
            path = save_markdown_report(result, RESULTS_DIR)
            console.print(f"[green]Markdown report saved: {path}[/green]")

    # Also save raw responses
    raw_path = RESULTS_DIR / f"{model.replace(':', '_')}_raw_responses.json"
    RESULTS_DIR.mkdir(parents=True, exist_ok=True)
    raw_data = []
    for r in responses:
        raw_data.append({
            "question_id": r.question_id,
            "raw_response": r.raw_response,
            "parsed_value": r.parsed_value,
            "refused": r.refused,
            "error": r.error,
        })
    with open(raw_path, "w") as f:
        json.dump(raw_data, f, indent=2)
    console.print(f"[dim]Raw responses saved: {raw_path}[/dim]")


if __name__ == "__main__":
    main()
