requirements.txt
openai>=1.12.0
anthropic>=0.25.0
click>=8.1.0
python-dotenv>=1.0.0
rich>=13.7.0
pytest>=8.0.0
